Move parser stuff into its own module
