+++ src/tags.rs
@@ -1,4 +1,4 @@
-use crate::{read_string, AsepriteParseError, Result};
+use crate::{parse::read_string, AsepriteParseError, Result};
 use byteorder::{LittleEndian, ReadBytesExt};
 use std::io::Cursor;
